\name{Rngstools-package}
\alias{Rngstools-package}
\alias{Rngstools}
\docType{package}
\title{
Utilities for dealing with next generation sequencing data
}
\description{
  Utilities for dealing with next generation sequencing data.

}
\details{
\tabular{ll}{
Package: \tab Rngstools\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-11-27\cr
License: \tab Artistic-2.0\cr
LazyLoad: \tab yes\cr
}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
Who wrote it

Maintainer: Per Unneberg <punneberg@gmail.com>
~~ Per Unneberg ~~
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
